Python/requirements.txt
pytest>=7.4.0
rich>=13.0.0
pyyaml>=6.0.1
pydantic>=2.7.0
# numpy is optional; tests do not require it strictly
numpy>=1.26.0; python_version>='3.11'
